requirements.txt
transformers>=4.40.0
torch>=2.1.0
flask>=3.0.0
pillow
accelerate
qwen-vl-utils
pydantic
torchvision